refactor: Header 분기처리하는 로직 커스텀훅으로 분리 #198
