+++ frontend/src/components/layout/Header/hooks/useRoutePath.ts
@@ -1,0 +1,24 @@
+import { useLocation, useParams } from 'react-router-dom';
+
+import { ROUTES } from '@/constants/routes';
+
+const useRoutePath = () => {
+  const location = useLocation();
+  const { roomId } = useParams();
+
+  const currentPath = {
+    isNicknamePage: location.pathname === ROUTES.nickname,
+    isReadyPage: location.pathname === ROUTES.ready(Number(roomId)),
+    isFinalResultPage: location.pathname === ROUTES.gameResult(Number(roomId)),
+    isRoundResultStatusPage: location.pathname === ROUTES.roundResultStatus(Number(roomId)),
+  };
+
+  return {
+    isNicknamePage: currentPath.isNicknamePage,
+    isReadyPage: currentPath.isReadyPage,
+    isRoundResultStatusPage: currentPath.isRoundResultStatusPage,
+    isFinalResultPage: currentPath.isFinalResultPage,
+  };
+};
+
+export default useRoutePath;
